% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantiles.R
\name{FindQuantile}
\alias{FindQuantile}
\title{Find Quantile of Ordinal Categorical Data.}
\usage{
FindQuantile(counts, q)
}
\arguments{
\item{counts}{Counts in each level.}

\item{q}{Quantile.}
}
\value{
Numeric interpolated quantile.
}
\description{
Find Quantile of Ordinal Categorical Data.
}
