% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{PermCounts}
\alias{PermCounts}
\title{Permute counts.}
\usage{
PermCounts(counts0, counts1)
}
\arguments{
\item{counts0}{Counts per level in the reference arm.}

\item{counts1}{Counts per level in the tretment arm.}
}
\value{
Levels by 2 matrix of permuted counts.
}
\description{
Permute counts.
}
