% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Index.R
\name{ProbIndex}
\alias{ProbIndex}
\title{Calculate the Probability Index.}
\usage{
ProbIndex(counts1, counts0)
}
\arguments{
\item{counts1}{Counts per level in the tretment arm.}

\item{counts0}{Counts per level in the reference arm.}
}
\value{
Numeric interpolated quantile.
}
\description{
Estimates the probability that the typical response in the 
treatment arm is greater than the typical response in the
reference arm, \eqn{P(Y_{1} > Y_{0})}.
}
