% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Index.R
\name{ProbIndexDiffP}
\alias{ProbIndexDiffP}
\title{Generate Permutation P-value for the Difference in in Probability Indices.}
\usage{
ProbIndexDiffP(counts0, counts1, B = 2000, q = 0.5)
}
\arguments{
\item{counts0}{Counts per level in the reference arm.}

\item{counts1}{Counts per level in the tretment arm.}

\item{B}{Bootstrap replicates.}

\item{q}{Quantile probability.}
}
\value{
Numeric vector containing the probability indices for arms 1 `Arm1` and 0 `Arm0`, 
the difference `Delta`, and the Pvalue.
}
\description{
Permutation p-value for the difference in probability indices:
\eqn{P(Y_{1} > Y_{0}) - P(Y_{0} > Y_{1})}.
}
