% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantiles.R
\name{QuantDiffCI}
\alias{QuantDiffCI}
\title{Generate Bootstrap Confidence Interval for the Difference in Quantiles.}
\usage{
QuantDiffCI(counts0, counts1, q = 0.5, B = 2000, alpha = 0.05)
}
\arguments{
\item{counts0}{Counts per level in the reference arm.}

\item{counts1}{Counts per level in the tretment arm.}

\item{q}{Quantile probability.}

\item{B}{Bootstrap replicates.}

\item{alpha}{Alpha level.}
}
\value{
Numeric vector containing the quantile probability `Quantile`, the
  estimated quantiles in arms 1 `Arm1` and 0 `Arm0`, the difference `Delta`,
  and the lower `L` and upper `U` confidence bounds.
}
\description{
Bootstrap confidence interval for the difference in quantiles.
}
