% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantiles.R
\name{QuantDiffP}
\alias{QuantDiffP}
\title{Generate Permutation P-value for the Difference in Quantiles.}
\usage{
QuantDiffP(counts0, counts1, B = 2000, q = 0.5)
}
\arguments{
\item{counts0}{Counts per level in the reference arm.}

\item{counts1}{Counts per level in the tretment arm.}

\item{B}{Bootstrap replicates.}

\item{q}{Quantile probability.}
}
\value{
Numeric vector containing the quantile probability `Quantile`, the
  estimated quantiles in arms 1 `Arm1` and 0 `Arm0`, the difference `Delta`,
  and the permutation p-value `P`.
}
\description{
Permutation p-value for the difference in quantiles.
}
